m44-9 move Nabvar related logic into its own component
