+++ src/components/navigation/navbar/Navbar.tsx
@@ -1,0 +1,42 @@
+import AppBar from '@mui/material/AppBar';
+import IconButton from '@mui/material/IconButton';
+import MenuIcon from '@mui/icons-material/Menu';
+import Toolbar from '@mui/material/Toolbar';
+import Typography from '@mui/material/Typography';
+import { useDispatch, useSelector } from 'react-redux';
+import { RootState } from '../../../state/store';
+import { toggleDrawer } from '../../../state/navigation/sidebarSlice';
+
+const Navbar = () => {
+  const dispatch = useDispatch();
+  const sidebarWidth = useSelector(
+    (state: RootState) => state.sidebar.sidebarWidth
+  );
+
+  return (
+    <AppBar
+      position="fixed"
+      sx={{
+        width: { sm: `calc(100% - ${sidebarWidth}px)` },
+        ml: { sm: `${sidebarWidth}px` },
+      }}
+    >
+      <Toolbar>
+        <IconButton
+          color="inherit"
+          aria-label="open drawer"
+          edge="start"
+          onClick={() => dispatch(toggleDrawer())}
+          sx={{ mr: 2, display: { sm: 'none' } }}
+        >
+          <MenuIcon />
+        </IconButton>
+        <Typography variant="h6" noWrap component="div">
+          Responsive drawer
+        </Typography>
+      </Toolbar>
+    </AppBar>
+  );
+};
+
+export default Navbar;
